Move FontFamily and FontWeight to new fonts module
I'd like to try adding FontStyle instead of a bool for italics,
which motivated a bit of cleanup.
