+++ piet/src/lib.rs
@@ -10,6 +10,7 @@ pub mod util;
 mod color;
 mod conv;
 mod error;
+mod font;
 mod gradient;
 mod null_renderer;
 mod render_context;
@@ -22,6 +23,7 @@ pub mod samples;
 pub use crate::color::*;
 pub use crate::conv::*;
 pub use crate::error::*;
+pub use crate::font::*;
 pub use crate::gradient::*;
 pub use crate::null_renderer::*;
 pub use crate::render_context::*;
